feat: Add structure to handle main bookshelves and add change shelf functionality
New array to handle the 3 main bookshelves in the project: Currently Reading, Want To Read and Read. New
React BookShelf component created. It was added the change shelf functionality in the onChange event for each
book. This functionality will modify the React state of the components and also will use the API to change the
shelf on the backend server